refactor: update import paths and update component names in UI files
